feat(dev): add a bundleable define entry point
Allow published design systems to bundle configuration helpers without retaining a runtime dependency on @pandacss/dev.
