+++ .changeset/add-dev-define-entrypoint.md
@@ -1,0 +1,6 @@
+---
+'@pandacss/dev': minor
+---
+
+Add an `@pandacss/dev/define` entry point for bundling configuration helpers into published libraries without including
+the full development package.
